try to use integer literal suffixes

+++ src/wlapplication.cc
@@ -22,7 +22,6 @@
 #ifndef _WIN32
 #include <csignal>
 #endif
-#include <cstddef>
 #include <cstdlib>
 #include <cstring>
 #include <iostream>
@@ -893,7 +892,7 @@ bool WLApplication::handle_key(bool down, const SDL_Keycode& keycode, const int 
 	if (matches_shortcut(KeyboardShortcut::kCommonScreenshot, keycode, modifiers)) {
 		if (g_fs->disk_space() < kMinimumDiskSpace) {
 			log_warn("Omitting screenshot because diskspace is lower than %lluMiB\n",
-			         kMinimumDiskSpace / (static_cast<uint64_t>(1024) * 1024));
+			         kMinimumDiskSpace / (1024ULL * 1024));
 		} else {
 			g_fs->ensure_directory_exists(kScreenshotsDir);
 			for (uint32_t nr = 0; nr < 10000; ++nr) {
@@ -1148,7 +1147,7 @@ bool WLApplication::init_settings() {
 
 	int64_t last_start = get_config_int("last_start", 0);
 	int64_t now = time(nullptr);
-	if (last_start + static_cast<int64_t>(12 * 60) * 60 < now ||
+	if (last_start + 12LL * 60LL * 60LL < now ||
 	    get_config_string("uuid", "").empty()) {
 		// First start of the game or not started for 12 hours. Create a (new) UUID.
 		// For the use of the UUID, see network/internet_gaming_protocol.h
